Fully functioning UI and Interactions with gatherables
The only blueprint is a BP_gather parent, witch you simply need to change the enum of to make it change between food, wood, and stone. we will change the placeholder cube within and add a static mesh. you can then set the static mesh to the actual mesh and material within the construction script.